align combobox and select components as closely as possible without merging
