+++ react/src/components/SelectableList.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styled from 'styled-components';
 
 const StyledList = styled.ul`
